requirements-test.txt
httpx
pydantic>=2,<3
pytest
pyyaml>=6,<7
